bugfix: return type propagation now works from attestation scripts

+++ src/action_flee_perform.c
@@ -6,7 +6,7 @@
  * action_flee_find_flee_position / action_flee_current_position_exposed.
  *
  * DEVIATIONS: (1) the DB prototype lists three params (actor_index, a2, a3) but the prologue reads only
- * r3 — a2/a3 are phantom (real signature is int action_flee_perform(int actor_index)); (2) the
+ * r3 — a2/a3 are phantom (real signature is uint8_t action_flee_perform(int actor_index)); (2) the
  * action_flee_find_flee_position call is rendered with a spurious third arg — its real signature is
  * (actor_index, flee_state_data*), matching the committed extern. The prop fields use prop_datum's
  * modeled names. */
@@ -27,8 +27,7 @@
 #include "headers/ai_information_data.h"
 extern int game_time_get(void);
 extern uint8_t action_flee_at_flee_position(int actor_index);
-extern uint8_t action_flee_current_position_exposed(uint16_t actor_index,
-        flee_state_data *flee_state_data);
+extern uint8_t action_flee_current_position_exposed(uint16_t actor_index, flee_state_data *flee_state_data);
 extern void action_flee_find_flee_position(int actor_index, flee_state_data *state_data);
 extern void actor_situation_update_target_status(uint16_t actor_index);
 extern void actor_situation_combat_status_update(uint16_t actor_index);
@@ -36,7 +35,7 @@ extern uint8_t unit_is_speaking(uint16_t unit_index);
 extern int unit_scream(int unit_index, int16_t scream_type);
 extern void ai_communication_event(int16_t communication_type, int subject_unit_index, int cause_unit_index, int16_t hostility, int16_t damage_type, int16_t information_type, ai_information_data *information_data);
 
-int action_flee_perform(int actor_index)
+uint8_t action_flee_perform(int actor_index)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
     flee_state_data *flee = &actor->state.action_data.___u0.flee;
