+++ src/action_vehicle_perform.c
@@ -48,7 +48,7 @@ extern uint8_t unit_enter_seat(int unit_index, int parent_unit_index, int16_t se
 extern uint8_t actor_move_halt(uint16_t actor_index);
 extern uint8_t actor_move_to_point(int actor_index, real_point3d *destination, int surface_index, int ignore_target_object_index);
 
-int action_vehicle_perform(int actor_index)
+uint8_t action_vehicle_perform(int actor_index)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
     vehicle_state_data *action_data = &actor->state.action_data.___u0.vehicle;
