+++ src/actor_action_can_stop_conversing.c
@@ -16,7 +16,7 @@
 #include "headers/blam_data_globals.h"
 
 
-uint8_t actor_action_can_stop_conversing(uint16_t actor_index)
+uint8_t actor_action_can_stop_conversing(int actor_index)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
     int conversation_index = actor->external_orders.conversation_index;
