+++ src/dsound_initialize.c
@@ -1,2 +1,3 @@
+#include <stdint.h>
 #include "headers/sound_preferences.h"
-int dsound_initialize(sound_preferences *preferences) { return 1; }
+uint8_t dsound_initialize(sound_preferences *preferences) { return 1; }
