+++ src/global_structure_bsp_index_get.c
@@ -3,7 +3,7 @@
 /* global_structure_bsp_index_get @0x83704048 — return the index of the currently-loaded structure BSP. */
 
 
-int global_structure_bsp_index_get(void)
+int16_t global_structure_bsp_index_get(void)
 {
-    return (uint16_t)global_structure_bsp_index;
+    return global_structure_bsp_index;
 }
