+++ src/hs_unit_can_see_flag.c
@@ -7,7 +7,7 @@
 
 extern uint8_t unit_can_see_point(int unit_index, const real_point3d *point, float theta);
 
-int hs_unit_can_see_flag(int unit_index, int16_t flag_index, float degrees)
+uint8_t hs_unit_can_see_flag(int unit_index, int16_t flag_index, float degrees)
 {
     if (!flag_index)
         return 0;
