+++ src/rasterizer_is_fullscreen.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 extern int _rasterizer_is_fullscreen(void);
 
-uint8_t rasterizer_is_fullscreen(void)
+int rasterizer_is_fullscreen(void)
 {
     return _rasterizer_is_fullscreen();
 }
